Refactor asClampedNanos(Duration) into own class
This utility should be used for Timeout as well as Retry
